requirement.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-background-tasks==1.2.5
django-ckeditor==6.1.0
django-compat==1.0.15
django-filter==21.1
django-js-asset==1.2.2
django-model-utils==4.2.0
django-taggit==1.5.1
django-widget-tweaks==1.4.9
ghasedak==0.1.9
idna==3.3
Pillow==8.4.0
pkg_resources==0.0.0
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
